Move code to the Finder
